Add signal handling for graceful shutdown
When the push client is managed using runit, it will receive a TERM
signal when the service is sent a stop or restart message. With this
patch, we trap the signal and call the client's shutdown method.

Also added is a handler for USR1 to manually trigger a client
reconfigure.