Verify encrypted socket is working
Since 0MQ doesn't return an error if there is a problem
with encryption, we'll only know it's not working if we
expect a message, but none comes through.  To ensure we
can identify this case, the server now always sends a
message on connection.  (Sometimes it is a new type: "ack",
sometimes it is the old "abort".)  If none comes within
a few seconds, we log an error, and retry the connection.