Delay reconfigures when a job is running.
The periodic reconfigurer can cause running jobs to terminate early;
this should mitigate but not outright fix the issue.

This is a minimal change fix that delays reconfiguration until any
running jobs have terminated. However this does have some race
conditions, in that:

1) The reconfigure is delayed by polling on idle; however we don't block
the acceptance of a new job immediately after reconfiguration, and that
job may be terminated

2) A new job may be started before the poller recognizes that the old
job has completed

A better fix would involve messing with the job_runner state machine and
making it easier to poll on, or blocking until we notify transition to
idle. However, the architecture of the system makes this a bit painful.