Fix ohai gem constraints
Chef 12.8.1 has a loose gem constraint, which will cause two versions of
ohai to be installed. This results in push-jobs throwing an error if a
newer version of Ohai than the one push-jobs uses is ever installed.

Until we upgrade Chef to a more modern version, have the gem use the
same constraints as the chef gem, and do not override the omnibus
version of the gem installed.

Signed-off-by: Tom Duffield <tom@chef.io>