Drop stdout and stderr from message if it exceeds MAX_BODY_SIZE
If a message's stdout and stderr are too large, the server can choke and
not properly mark the job as complete. Instead of leaving the job in an
unknown state, drop the stdout and stderr keys from the message and log
a warning.

Also add libzmq3-dev to travis so it can compile ffi-rzmq.

Signed-off-by: Nathan L Smith <smith@chef.io>