Wrap ZMQ request in timeout
As noted in #123, the push jobs client will sometimes hang on the
send_string call when the push jobs server is restarted (or is replaced
by another PJ server).  After adding this timeout, I have been able to
successfully restart and replace the PJ server without the client
hanging consistently.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>