Quick hack at a reconfigure strategy
This modifies the periodic reconfigure loop to allow us to alter the
reconfigure timeout, and makes packet send timeouts advance the reconfigure deadline.

It seems probable if we are timing out in zeromq, we have a broken
connection (as if we failed to recieve heartbeats) and should treat it
the same way.

Signed-off-by: Mark Anderson <mark@chef.io>