Adding reconfigure to rescue
Brought in Mark's changes from PR #144 with some modifications.

Added an instance method to PushyClient to expose the
update_reconfigure_deadline method through the client.  Also added the
client as a param to the send_signed_message method, so we can access
it for both the node_name and update_reconfigure_deadline methods.

One issue I noticed was that if the get_config method failed after 5
retries, the client would hang until the next reconfigure timeout (3600
seconds).  Added a begin/rescue around the API call to download the
config to make use of our new update_reconfigure_deadline method to
schedule another reconfig in 120 seconds (total guess on the time,
possibly should be longer?).  I raised an error after setting the next
reconfigure, which may or may not be the most sane thing to do.  It
basically replicates the current behavior (rest.get(resource) throws an
exception if the config is not downloaded successfully), but schedules
a reconfig before doing so.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>