Optimize `get_config` by eliminating extensive retry behavior
This change attempts to address an edge case scenario where the push client
service becomes wedged, presumably after or during a network outage.

Customer logs show the typical retry behavior in `Chef::HTTP` when the `reconfigure`
block executes the `get_config` code path. Afterwards, there seems to be
an infinite deadlock with no further log messages from within the `reconfigure`.

This probably means that the `@reconfigure_lock.synchronize do` block held
onto the mutex lock, began executing the block code but then never released
the lock.

When diagnosing what was happening within that block I noticed the `get_config`
could lead to a behavior of extensive time being spent in HTTP retries during
a network outage. Depending on the situation, if the requests were timing out,
the default value for timeouts is [300](https://github.com/chef/chef/blob/master/chef-config/lib/chef-config/config.rb#L432) seconds.
Combined with the default 5 retries, that can be an extensive time for the
mutex to hold a lock. Meanwhile other threads executing have a `reconfigure`
triggered but can never obtain a lock on the mutex.

The commit in this PR addresses some rubocop warnings and introduces a
different behavior to `get_config` by disabling the typical `Chef::HTTP` retry
pattern when being executed during a `reconfigure`.

The goal it to get `get_config` to fail fast and not retry. There already
is an effective retry loop implemented by the `reconfigure` and
`update_reconfigure_deadline` methods.

The following was performed on Windows 2012R2 and Centos 7

1. Apply changes to `get_config`
2. Launch push client process
3. Introduce network abnormalities communicating to Chef Server with [netem](https://wiki.linuxfoundation.org/networking/netem)
4. Observe the mutex lock, `get_config` fail fast, and the subsequent lock release.
5. Remove network outage and observe hearbeats resume as expected.

- [x] Code actually executed?
- [x] Vetting performed (unit tests, lint, etc.)?

Signed-off-by: Jeremy J. Miller <jm@chef.io>